push locking into rate structure and make them spin when required
To limit possible contention when updating bitrates for source and global, move
away from a mutex lock and create spin lock per rate structure. The functions are
re-worked to only lock when required so should scale much better.

having these locks here allows us to get rid of the recent spin locks in source_t
and global preventing memory allocation delaying the locking and goes some way to
allow scaling in other areas.

+++ src/util.h
@@ -61,7 +61,7 @@ char *util_conv_string (const char *string, const char *in_charset, const char *
 
 struct rate_calc *rate_setup (unsigned int samples, unsigned int ssec);
 void rate_add (struct rate_calc *calc, long value, uint64_t t);
-float rate_avg (struct rate_calc *calc);
+long rate_avg (struct rate_calc *calc);
 void rate_free (struct rate_calc *calc);
 void rate_reduce (struct rate_calc *calc, unsigned int range);
